Added standard curls and created new helper for other functionalities
