+++ app/code/Mageseller/DriveFx/etc/di.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <preference for="Magento\Framework\HTTP\Client\Curl" type="Mageseller\DriveFx\HTTP\Client\Curl"/>
+</config>
